+++ components/mobileMenu/MobileMenu.tsx
@@ -1,91 +1,8 @@
-import { FC, useEffect, useLayoutEffect, useState } from "react";
-import styled from "styled-components";
+import { FC, useEffect, useLayoutEffect } from "react";
 import useEventListener from "../../lib/hooks/useEventListener";
+import useMediaQuery from "../../lib/hooks/useMediaQuery";
 import VaultNavigation from "../vaultNavigation";
-
-function useMediaQuery(query: string): boolean {
-  const getMatches = (query: string): boolean => {
-    // Prevents SSR issues
-    if (typeof window !== "undefined") {
-      return window.matchMedia(query).matches;
-    }
-    return false;
-  };
-
-  const [matches, setMatches] = useState<boolean>(getMatches(query));
-
-  function handleChange() {
-    setMatches(getMatches(query));
-  }
-
-  useEffect(() => {
-    const matchMedia = window.matchMedia(query);
-
-    // Triggered at the first client-side load and if query changes
-    handleChange();
-
-    // Listen matchMedia
-    if (matchMedia.addListener) {
-      matchMedia.addListener(handleChange);
-    } else {
-      matchMedia.addEventListener("change", handleChange);
-    }
-
-    return () => {
-      if (matchMedia.removeListener) {
-        matchMedia.removeListener(handleChange);
-      } else {
-        matchMedia.removeEventListener("change", handleChange);
-      }
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [query]);
-
-  return matches;
-}
-
-const Wrapper = styled.div`
-  background-color: rgba(0, 0, 0, 0.2);
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  z-index: 99999999;
-`;
-
-const Overlay = styled.div`
-  background-color: black;
-  width: 100px;
-  height: 100px;
-`;
-
-const Menu = styled.div`
-  background-color: #fff;
-  width: 420px;
-  height: 100vh;
-
-  .top-area {
-    align-items: center;
-    display: flex;
-    height: 84px;
-    padding: 0 16px;
-  }
-`;
-
-const CloseButton = styled.button`
-  background-color: transparent;
-  border: none;
-  border-radius: 99999px;
-  color: var(--gray-500);
-  padding: 4px;
-
-  svg {
-    display: block;
-    height: 32px;
-    width: 32px;
-  }
-`;
+import { CloseButton, Menu, Wrapper } from "./elements";
 
 interface Props {
   onClose: () => void;
