+++ components/navbar/NavBar.tsx
@@ -2,7 +2,6 @@
 import dynamic from "next/dynamic";
 import React, { FC } from "react";
 import styled from "styled-components";
-import MobileMenu from "../mobileMenu";
 
 const Actions = dynamic(() => import("./NavBarActions"), {
   ssr: false,
@@ -51,28 +50,25 @@ const Wrapper = styled.nav`
 
 const NavBar: FC = () => {
   return (
-    <React.Fragment>
-      <Wrapper>
-        <div className="inner">
-          <img className="logo" alt="Valorem logo" src="/logo.png" />
-          <div className="mobile-logo">
-            <svg
-              width="70"
-              height="58"
-              viewBox="0 0 70 58"
-              xmlns="http://www.w3.org/2000/svg"
-            >
-              <path
-                d="M69.3577 14.5031H29.7265L39.6312 0H0L19.8156 29L29.7265 14.5031L39.6312 29H19.8156H0L19.8156 58L39.6312 29L49.5421 43.5031L69.3577 14.5031Z"
-                fill="currentColor"
-              />
-            </svg>
-          </div>
-          <Actions />
+    <Wrapper>
+      <div className="inner">
+        <img className="logo" alt="Valorem logo" src="/logo.png" />
+        <div className="mobile-logo">
+          <svg
+            width="70"
+            height="58"
+            viewBox="0 0 70 58"
+            xmlns="http://www.w3.org/2000/svg"
+          >
+            <path
+              d="M69.3577 14.5031H29.7265L39.6312 0H0L19.8156 29L29.7265 14.5031L39.6312 29H19.8156H0L19.8156 58L39.6312 29L49.5421 43.5031L69.3577 14.5031Z"
+              fill="currentColor"
+            />
+          </svg>
         </div>
-      </Wrapper>
-      {/* <MobileMenu /> */}
-    </React.Fragment>
+        <Actions />
+      </div>
+    </Wrapper>
   );
 };
